Extrai um caminho pra uma classe wrapper
