+++ Assets/Scripts/Tiles/DoorTileType.cs
@@ -5,7 +5,6 @@ namespace DonBigo
 {
     public class DoorTileType : WallTileType, IRoomEntranceMarker
     {
-        //TODO armazena informações sobre estado de abertura, trancado, mudança de sprite, etc
         [SerializeField] private WallTileType inactiveSubstitute;
 
         public void SetInactive(GameGrid grid, Tilemap tilemap, Vector2Int pos)
